requirements.txt
flask>=3.0.3
flask-cors>=4.0.1
opencv-python>=4.10.0.84
ultralytics>=8.3.0
torch
torchvision
deep-sort-realtime>=1.3.2
streamlit>=1.37.1
python-dotenv>=1.0.1
requests>=2.32.3
google-generativeai>=0.7.2
playsound==1.2.2
numpy
Pillow
scipy
